Update README.md
Problem Definition: 
Many hotels and restaurants around the world, especially in developing countries are still struggling with un-automated billing systems. This way, such restaurants lose a lot of customers to their competitors as a result of improper billing systems. In addition, such systems lead to the loss of tourist clients from both abroad and within the country due to the lack of digital platforms to access and order various menus. 
The purpose of this hotel billing system is to fully automate the process of ordering various food items on the menu. 

System Design: 
This project is designed in the graphical user interface (GUI) application 
Uses a tkinter module to display and create a user-friendly interface.
The project is designed and well-protected through the Python coding structure and contains automated features that make transactions faster and more convenient. 
The system is fully automated where the user can order items of choice and the bills are automatically generated.
The system is user friendly and one can fully access this hotel billing system without entering the user login details.